volatile-binds: use VOLATILE_BINDS_append instead of VOLATILE_BINDS +=
For some reason, the latter was breaking having an empty or non-existent
VOLATILE_BINDS_append in mel.conf.

Also use the bitbake variables instead of hardcoding the paths.

Signed-off-by: Benjamin Walsh <Benjamin_Walsh@mentor.com>

This change is similar to meta-mentor-bsp commit id 7fc6936 and is the
continuation of the previous commit.

JIRA: http://jira.alm.mentorg.com:8080/browse/MEIF-1352

Signed-off-by: Abdur Rehman <abdur_rehman@mentor.com>